Moved getGroupedMemoryReport to class and fixed issue with buffer colours
